requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
boto3==1.26.112
botocore==1.29.112
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
flake8==6.0.0
gunicorn==20.1.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
